Began implementing the morale system and refactoring of Turnover code
Made extensive changes to the organization and implementation of the morale and turnover rules, including the addition of a step size option and a theft function. Also renamed and updated various methods and properties for improved code clarity and readability.

+++ MekHQ/src/mekhq/campaign/io/CampaignXmlParser.java
@@ -276,7 +276,7 @@ public class CampaignXmlParser {
                 } else if (xn.equalsIgnoreCase("retirementDefectionTracker")) {
                     retVal.setRetirementDefectionTracker(RetirementDefectionTracker.generateInstanceFromXML(wn, retVal));
                 } else if (xn.equalsIgnoreCase("Morale")) {
-                    retVal.setMorale(Integer.parseInt(wn.getTextContent()));
+                    retVal.setMorale(Double.parseDouble(wn.getTextContent()));
                 } else if (xn.equalsIgnoreCase("shipSearchStart")) {
                     retVal.setShipSearchStart(MHQXMLUtility.parseDate(wn.getTextContent().trim()));
                 } else if (xn.equalsIgnoreCase("shipSearchType")) {
